wip: lua floating window/terminal manager prototype
Adds floating window functionality for terminal and windows (written in
Lua). This is still a work-in-progress, and there are a few hiccups
which still need to be solved.

The aim is to create a 'standardised' layout of floating windows which
can be configured using pre-determined/personal layout options (e.g.
opening a running script as a central floating terminal, whilst opening
REPLs in the top-right corner of the editor) with some nice features
such as sending visual selections to a floating terminal, opening active
buffers in a floating window etc.