checkpoint: port `.vimrc` from VimScript to Lua
This has been worked on for the past couple of months and I had
forgotten to commit my changes over time, so a large 'checkpoint' commit
has been added now that I have converted approximately 90% of my
original VimScript `.vimrc` to Lua.

There are still enhancements to be made over time, but these should
hopefully remain as small commits. The next steps for my Neovim
configuration are to:

- Port all `ftplugin` files to Lua.

- Port `auto_directory` utility to Lua.

- Modularise the LSP setup.