update: refactor `ftplugin/lua.lua`, and add documentation
The functions inside this file are taken from all over the Internet, and
it was very difficult to piece together everything for the purpose of
using `gf` to navigate Lua files due to verbose function naming from
other people. I decided to refactor this code and clean up function
names so it easier to understand.

I will definitely forget how each function in this file works, so I
decided to heavily document it so (a) I can understand it in 6 months
time, and (b) so other people can understand it when viewing.