fix: overwrite markdown conceal for code blocks
A change to the markdown treesitter parser added code blocks to it's
default conceal rules. These are set unconditionally so there is no way
to easily enable/disable individual conceals. The code in this commit
temporarily addresses [this issue](https://github.com/nvim-treesitter/nvim-treesitter/issues/2825),
and this code will be removed if there is a fix for this issue.