feat: run python file in split with `:Run`
This was a feature that I had in my Vim configuration a while back, but
I never ported it to Neovim because of their floating window
architecture. I might still use the floating window in the future, but
for now I quite like being able to quickly run a script in a vertical
split.